requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pandas
streamlit
python-multipart
requests
tqdm
pytest
python-dotenv
